pyproject.toml
[tool.poetry]
name = "pypi-server"
version = "0.1.0"
description = "Source for Seebak package"
authors = ["Luis Díaz <luigirdf@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
aws-cdk-lib = "2.1.0"
constructs = ">=10.0.0,<11.0.0"

[tool.poetry.dev-dependencies]
pytest = "6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
